pyproject.toml
[tool.poetry]
name = "bulletin-board"
version = "0.1.0"
description = ""
authors = ["skip"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.6.0"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.30.0"

[tool.poetry.group.lint.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
